// src/server/PixelBattles.Server.Web/Client/PixelBattle/IApiClient.ts
import { IHubClient } from "./IHubClient";

export interface IApiClient {
    createHubClient(gameId: string, url: string, token: string): Promise<IHubClient>;
    getBattleInfo(battleId: string): Promise<Battle>; 
}

export interface Battle {
    battleId: string;
    name: string;
    description: string;
}

export interface Game {
    gameId: string;
    height: number;
    width: number;
    cooldown: number;
    startDateUTC: string;
    endDateUTC: string;
}